Add rviz reset, reduce ROS publish rate in sim, fix robot position in occupancy grid,

+++ src/taser/navigation/environments/occupancy_grid.py
@@ -32,10 +32,10 @@ class OccupancyGrid:
             )
             self._grid = np.zeros((y_shape, x_shape))
 
-    def set(self, region: tuple[int, int, int, int], value: float) -> None:
+    def set(self, region: tuple[float, float, float, float], value: float) -> None:
         """Set the value of a region in the occupancy grid.
         Args:
-            region (tuple[int, int, int, int]): The region to set in the format (x_min, x_max, y_min, y_max).
+            region (tuple[float, float, float, float]): The region to set in the format (x_min, x_max, y_min, y_max).
             value (float): The value to set the region to (0 for free, 1 for occupied).
         """
         x_min, x_max, y_min, y_max = region
